write.sav

Function to write an SPSS sav or zsav file from a data.frame().

write.sav(
  dat,
  filepath,
  label,
  add.rownames = FALSE,
  compress = FALSE,
  convert.dates = TRUE,
  tz = "GMT",
  debug = FALSE,
  is_zsav = FALSE,
  disppar
)

Arguments

dat

data.frame a data.frame to store as SPSS file.

filepath

string full path where and how this file should be stored

label

character if any provided this must be a vector of labels. It must be of size ncol(dat)

add.rownames

logical If TRUE, a new variable rownames will be added to the sav-file.

compress

logical should compression be used. If TRUE some integers will be stored more efficiently. Everything will be stored in chunks of 8 chars. Reduces memory size of sav-file.

convert.dates

logical should dates be converted to SPSS format.

tz

character The name of the timezone convert.dates will use.

debug

logical print debug information.

is_zsav

logical explicitly create a zsav file. If the file ending zsav is used, this is selected as default.

disppar

optional display parameter matrix. Needs documentation.

Value

write.sav returns nothing

Details

Writing of strings longer than 255 chars is not provided.